- moved the repository classes to a separate repository package inside the phoenicis-apps project - added toString methods for the most often used repository types - added equals and hashCode methods for the NullRepository - removed the filter classes from the phoenicis-apps project
