--- phoenicis-apps/src/main/java/org/phoenicis/apps/CombinedAppsFilter.java
@@ -1,56 +1,0 @@
-/*
- * Copyright (C) 2015-2017 PÂRIS Quentin
- *
- * This program is free software; you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation; either version 2 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License along
- * with this program; if not, write to the Free Software Foundation, Inc.,
- * 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
- */
-
-package org.phoenicis.apps;
-
-
-import org.phoenicis.apps.dto.ApplicationDTO;
-
-import java.util.HashSet;
-import java.util.Set;
-
-public class CombinedAppsFilter implements AppsFilter{
-    private final Set<AppsFilter> filters = new HashSet<>();
-
-    public CombinedAppsFilter() {
-    }
-
-    public void remove(AppsFilter filter) {
-        filters.remove(filter);
-    }
-
-    public void add(AppsFilter filter) {
-        filters.add(filter);
-    }
-
-    public void clear() { filters.clear(); }
-
-    public boolean applies(ApplicationDTO applicationDTO) {
-        if (filters.isEmpty()) {
-            return true;
-        }
-
-        boolean applies = false;
-        for (AppsFilter filter : filters) {
-            if (filter.applies(applicationDTO)) {
-                return true;
-            }
-        }
-        return applies;
-    }
-}
