+++ phoenicis-apps/src/main/java/org/phoenicis/apps/repository/Repository.java
@@ -16,7 +16,7 @@
  * 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
  */
 
-package org.phoenicis.apps;
+package org.phoenicis.apps.repository;
 
 import org.phoenicis.apps.dto.ApplicationDTO;
 import org.phoenicis.apps.dto.CategoryDTO;
@@ -55,8 +55,6 @@ public interface Repository {
 
         return null;
     }
-
-    default void setFilter(CombinedAppsFilter filter) {}
 
     default void getScript(List<String> path, Consumer<ScriptDTO> callback, Consumer<Exception> errorCallback) {
         callback.accept(getScript(path));
