+++ phoenicis-tests/src/main/java/org/phoenicis/tests/MockedRepository.java
@@ -18,9 +18,9 @@
 
 package org.phoenicis.tests;
 
-import org.phoenicis.apps.Repository;
+import org.phoenicis.apps.repository.Repository;
 import org.phoenicis.entities.OperatingSystem;
-import org.phoenicis.apps.TeeRepository;
+import org.phoenicis.apps.repository.TeeRepository;
 import org.phoenicis.apps.dto.ApplicationDTO;
 import org.phoenicis.apps.dto.CategoryDTO;
 import org.phoenicis.apps.dto.ScriptDTO;
