+++ phoenicis-tests/src/main/java/org/phoenicis/tests/PhoenicisTestsConfiguration.java
@@ -19,8 +19,8 @@
 package org.phoenicis.tests;
 
 import org.phoenicis.apps.AppsConfiguration;
-import org.phoenicis.apps.NullRepository;
-import org.phoenicis.apps.Repository;
+import org.phoenicis.apps.repository.NullRepository;
+import org.phoenicis.apps.repository.Repository;
 import org.phoenicis.configuration.PhoenicisGlobalConfiguration;
 import org.phoenicis.multithreading.MultithreadingConfiguration;
 import org.phoenicis.scripts.ScriptsConfiguration;
